+++ tests/auto/qml/qqmljsscope/tst_qqmljsscope.cpp
@@ -247,7 +247,8 @@ void tst_qqmljsscope::allTypesAvailable()
         };
 
         QQmlJSImporter importer { importPaths, /* resource file mapper */ nullptr };
-        const auto imported = importer.importModule(u"QtQml"_s, QQmlJS::PrecedenceValues::Default);
+        const auto imported = importer.importModule(u"QtQml"_s,
+                                                    quint8(QQmlJS::PrecedenceValues::Default));
         QCOMPARE(imported.contextualTypes().context(), QQmlJS::ContextualTypes::QML);
         const auto types = imported.types();
         QVERIFY(types.contains(u"$internal$.QObject"_s));
